feature show team recommend for champ

+++ data/src/main/java/com/tft_mvvm/data/features/champs/repository/RepoRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.tft_mvvm.data.features.champs.repository
 
+import android.util.Log
 import com.example.common_jvm.function.Either
 import com.example.common_jvm.function.Either.Companion.runSuspendWithCatchError
 import com.tft_mvvm.data.exception_interceptor.RemoteExceptionInterceptor
@@ -10,7 +11,9 @@ import com.tft_mvvm.data.local.database.ClassAndOriginDAO
 import com.tft_mvvm.data.local.database.ItemDAO
 import com.tft_mvvm.data.local.database.TeamDAO
 import com.tft_mvvm.data.local.model.ChampListDBO
-import com.tft_mvvm.domain.features.model.*
+import com.tft_mvvm.domain.features.model.ChampListEntity
+import com.tft_mvvm.domain.features.model.TeamBuilderListEntity
+import com.tft_mvvm.domain.features.model.TeamListEntity
 import com.tft_mvvm.domain.features.repository.RepoRepository
 
 class RepoRepositoryImpl(
@@ -33,6 +36,7 @@ class RepoRepositoryImpl(
 
     private fun createChamp(
         champDBO: ChampListDBO.ChampDBO,
+        threeStart: Boolean,
         listItem: MutableList<ChampListEntity.Champ.Item>
     ): ChampListEntity.Champ {
         return ChampListEntity.Champ(
@@ -47,22 +51,23 @@ class RepoRepositoryImpl(
             linkSkillAvatar = champDBO.linkSkillAvatar,
             rankChamp = champDBO.rankChamp,
             suitableItem = listItem,
+            threeStar = threeStart,
             name = champDBO.name
         )
     }
 
     override suspend fun getChamps(isForceLoadData: Boolean) =
         runSuspendWithCatchError(listOf(remoteExceptionInterceptor)) {
-            val listChampDBO = mutableListOf<ChampListDBO.ChampDBO>()
+            var listChampDBO = champDAO.getAllChamp()
             val listChampEntity = mutableListOf<ChampListEntity.Champ>()
             if (champDAO.getAllChamp().isNullOrEmpty() || isForceLoadData) {
-                if (isForceLoadData) {
-                    champDAO.deleteAllChampTable()
-                }
                 val listChampResponse = apiService.getChampList()
                 val listChampDbo = champDaoEntityMapper.map(listChampResponse)
+                if (isForceLoadData && listChampDbo.champDBOs.isNotEmpty()) {
+                    champDAO.deleteAllChampTable()
+                }
                 champDAO.insertChamps(listChampDbo.champDBOs)
-                listChampDBO.addAll(champDAO.getAllChamp())
+                listChampDBO = champDAO.getAllChamp()
             }
 
             if (itemDAO.getAllItem().isNullOrEmpty()) {
@@ -80,7 +85,7 @@ class RepoRepositoryImpl(
                 }
 
                 listChampEntity.add(
-                    createChamp(i, listItem)
+                    createChamp(i, false, listItem)
                 )
             }
             listChampEntity.size
@@ -118,11 +123,11 @@ class RepoRepositoryImpl(
                     )
                 )
             if (dbTeamListEntity.teams.isNullOrEmpty() || isForceLoadData) {
-                if (isForceLoadData) {
-                    teamDAO.deleteAllTeam()
-                }
                 val teamListResponse = apiService.getTeamList()
                 val teamListDBO = teamDaoEntityMapper.map(teamListResponse)
+                if (isForceLoadData && teamListDBO.teamDBOs.isNotEmpty()) {
+                    teamDAO.deleteAllTeam()
+                }
                 teamDAO.insertTeam(teamListDBO.teamDBOs)
                 dbTeamListEntity =
                     TeamListEntity(
@@ -145,13 +150,30 @@ class RepoRepositoryImpl(
                 val listIdChampCommon = mutableListOf<String>()
                 listIdChampCommon.addAll(team.listIdChamp)
                 for (position in listChampMainDbo.indices) {
+                    var check = 0
+                    for (idChampThreeStart in team.listIdChampThreeStar) {
+                        if (listChampMainDbo[position].id == idChampThreeStart) {
+                            check++
+                            Log.d("phuc", "$check")
+                        }
+                    }
                     val listItem = mutableListOf<ChampListEntity.Champ.Item>()
                     val listIdItem = team.listIdSuitable[position].split(",")
                     for (idItem in listIdItem) {
                         val item = itemListMapper.map(itemDAO.getItemById(idItem))
                         listItem.add(item)
                     }
-                    listChampEntity.add(createChamp(listChampMainDbo[position], listItem))
+                    if (check > 0) {
+                        listChampEntity.add(createChamp(listChampMainDbo[position], true, listItem))
+                    } else {
+                        listChampEntity.add(
+                            createChamp(
+                                listChampMainDbo[position],
+                                false,
+                                listItem
+                            )
+                        )
+                    }
                     val index = mutableListOf<Int>()
                     for (positionIdChampCommon in listIdChampCommon.indices) {
                         if (listChampMainDbo[position].id == listIdChampCommon[positionIdChampCommon]) {
@@ -162,9 +184,34 @@ class RepoRepositoryImpl(
                         listIdChampCommon.removeAt(p)
                     }
                 }
-                val listChampCommon =
-                    champListMapper.mapList(champDAO.getListChampByTeam(listIdChampCommon))
-                listChampEntity.addAll(listChampCommon)
+
+                for (idChamp in listIdChampCommon) {
+                    var check = 0
+                    for (idChampThreeStart in team.listIdChampThreeStar) {
+                        if (idChamp == idChampThreeStart) {
+                            check++
+                        }
+                    }
+                    if (check > 0) {
+                        listChampEntity.add(
+                            createChamp(
+                                champDAO.getChampById(idChamp),
+                                true,
+                                mutableListOf()
+                            )
+                        )
+                    } else {
+                        listChampEntity.add(
+                            createChamp(
+                                champDAO.getChampById(idChamp),
+                                false,
+                                mutableListOf()
+                            )
+                        )
+                    }
+                }
+                listChampEntity.sortBy { it.name }
+                listChampEntity.sortBy { it.cost }
                 listTeamBuilder.add(
                     TeamBuilderListEntity.TeamsBuilder(
                         team.nameTeam,
@@ -186,7 +233,6 @@ class RepoRepositoryImpl(
     ) =
         runSuspendWithCatchError(listOf(remoteExceptionInterceptor)) {
             if (classAndOriginDAO.getAllClassAndOrigin().isNullOrEmpty() || isForceLoadData) {
-                classAndOriginDAO.deleteAllClassAndOrinTable()
                 val classAndOriginResponses = apiService.getClassAndOriginList()
                 val classAndOriginDBO = classAndOriginDaoEntityMapper.map(classAndOriginResponses)
                 classAndOriginDAO.insertClassAndOrigin(classAndOriginDBO.classAndOrigins)
@@ -231,21 +277,117 @@ class RepoRepositoryImpl(
                     listItem.add(itemListMapper.map(itemDAO.getItemById(i)))
                 }
             }
-            val dbResult = ChampListEntity.Champ(
-                id = champDBO.id,
-                cost = champDBO.cost,
-                linkImg = champDBO.linkImg,
-                activated = champDBO.activated,
-                linkChampCover = champDBO.linkChampCover,
-                classs = champDBO.classs,
-                origin = champDBO.origin,
-                skillName = champDBO.skillName,
-                linkSkillAvatar = champDBO.linkSkillAvatar,
-                rankChamp = champDBO.rankChamp,
-                suitableItem = listItem,
-                name = champDBO.name
-            )
-            return@runSuspendWithCatchError Either.Success(dbResult)
+            val champ = createChamp(champDBO, false, listItem)
+            return@runSuspendWithCatchError Either.Success(champ)
+        }
+
+    override suspend fun getTeamRecommendForChamp(id: String) =
+        runSuspendWithCatchError(listOf(remoteExceptionInterceptor)) {
+            var teamListEntity =
+                TeamListEntity(
+                    teams = teamListMapper.mapList(
+                        teamDAO.getAllTeam()
+                    )
+                )
+            if (teamListEntity.teams.isNullOrEmpty()) {
+                val listItemResponse = apiService.getItemListResponse()
+                val listItemDBO = itemDaoEntityMapper.map(listItemResponse)
+                itemDAO.insertItems(listItemDBO.items)
+                teamListEntity =
+                    TeamListEntity(
+                        teams = teamListMapper.mapList(
+                            teamDAO.getAllTeam()
+                        )
+                    )
+            }
+            val dbTeamListEntityResult = mutableListOf<TeamListEntity.Team>()
+            for (team in teamListEntity.teams) {
+
+                for (idChamp in team.listIdChamp) {
+                    if (id == idChamp) {
+                        dbTeamListEntityResult.add(team)
+                    }
+                }
+            }
+            val listTeamBuilder: MutableList<TeamBuilderListEntity.TeamsBuilder> = mutableListOf()
+            dbTeamListEntityResult.size
+            dbTeamListEntityResult.forEach { team ->
+                val listIdThreeStart = team.listIdChampThreeStar
+                val listChampEntity = mutableListOf<ChampListEntity.Champ>()
+                val listChampMainDbo = champDAO.getListChampByTeam(team.listIdChampMain)
+                val listIdChampCommon = mutableListOf<String>()
+                listIdChampCommon.addAll(team.listIdChamp)
+                for (position in listChampMainDbo.indices) {
+                    var check = 0
+                    for (idChampThreeStart in listIdThreeStart) {
+                        if (listChampMainDbo[position].id == idChampThreeStart) {
+                            check++
+                        }
+                    }
+                    val listItem = mutableListOf<ChampListEntity.Champ.Item>()
+                    val listIdItem = team.listIdSuitable[position].split(",")
+                    for (idItem in listIdItem) {
+                        val item = itemListMapper.map(itemDAO.getItemById(idItem))
+                        listItem.add(item)
+                    }
+                    if (check > 0) {
+                        listChampEntity.add(createChamp(listChampMainDbo[position], true, listItem))
+                    } else {
+                        listChampEntity.add(
+                            createChamp(
+                                listChampMainDbo[position],
+                                false,
+                                listItem
+                            )
+                        )
+                    }
+                    val index = mutableListOf<Int>()
+                    for (positionIdChampCommon in listIdChampCommon.indices) {
+                        if (listChampMainDbo[position].id == listIdChampCommon[positionIdChampCommon]) {
+                            index.add(positionIdChampCommon)
+                        }
+                    }
+                    for (p in index) {
+                        listIdChampCommon.removeAt(p)
+                    }
+                }
+                for (idChamp in listIdChampCommon) {
+                    var check = 0
+                    for (idChampThreeStart in team.listIdChampThreeStar) {
+                        if (idChamp == idChampThreeStart) {
+                            check++
+                        }
+                    }
+                    if (check > 0) {
+                        listChampEntity.add(
+                            createChamp(
+                                champDAO.getChampById(idChamp),
+                                true,
+                                mutableListOf()
+                            )
+                        )
+                    } else {
+                        listChampEntity.add(
+                            createChamp(
+                                champDAO.getChampById(idChamp),
+                                false,
+                                mutableListOf()
+                            )
+                        )
+                    }
+                }
+                listChampEntity.sortBy { it.name }
+                listChampEntity.sortBy { it.cost }
+                listTeamBuilder.add(
+                    TeamBuilderListEntity.TeamsBuilder(
+                        team.nameTeam,
+                        ChampListEntity(champs = listChampEntity)
+                    )
+                )
+
+            }
+            Log.d("phuc", "$listTeamBuilder")
+            return@runSuspendWithCatchError Either.Success(TeamBuilderListEntity(teamBuilders = listTeamBuilder))
         }
 
 }
